setup user auth

+++ app/Http/Controllers/PostsController.php
@@ -3,15 +3,26 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 
 class PostsController extends Controller
 {
     /**
+     * Prevents guests from accessing posts
+     * Exception applied to index and show
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show'] ]);
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
+    
     public function index()
     {
         // $posts = Post::all(); 
@@ -44,15 +55,34 @@ class PostsController extends Controller
         //validation 
         $this->validate($request, [
             'title' => 'required', //must have a title to submit
-            'body' => 'required'   //must have a body to submit
+            'body' => 'required',   //must have a body to submit
+            'cover_image' => 'image|nullable|max:1999' //set max of 2MB (apache server default size 2MB)
         ]);
 
+        //Handle file upload
+        if($request->hasFile('cover_image')) {
+            //Get filename w/ extension
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //Get filename only
+            $fileName = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            //Get ext only
+            $ext = $request->file('cover_image')->getClientOriginalExtension();
+
+            //Filename to store
+            $fileNameToStore = $fileName.'_'.time().'.'.$ext;  //prevents override of img with same name as previous img
+            //Upload img
+            $path = $request->file('cover_image')->storeAs('public/cover_image', $fileNameToStore);
+        } else {
+            $fileNameToStore = 'noimage.jpg';
+        }
+
         //create post
-        $post = new Post;  //create new post var
-        $post->title = $request->input('title');
-        $post->body = $request->input('body');
-        $post->user_id = auth()->user()->id;
-        $post->save();
+        $p = new Post;  //create new post var
+        $p->title = $request->input('title');
+        $p->body = $request->input('body');
+        $p->user_id = auth()->user()->id;
+        $p->cover_image = $fileNameToStore;
+        $p->save();
 
         return redirect('/posts')->with('success', 'Post Created');
     }
@@ -78,6 +108,13 @@ class PostsController extends Controller
     public function edit($id)
     {
         $p = Post::find($id);
+
+        //check for correct user so you cannot manually add edit to url
+        if(auth()->user()->id !== $p->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
+
         return view('posts/edit')->with('p', $p);
     }
 
@@ -93,13 +130,30 @@ class PostsController extends Controller
         //validation 
         $this->validate($request, [
             'title' => 'required', //must have a title to submit
-            'body' => 'required'   //must have a body to submit
+            'body' => 'required',   //must have a body to submit
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+        
+        //Handle file upload
+        if($request->hasFile('cover_image')) {
+            //Get filename w/ extension
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //Get filename only
+            $fileName = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            //Get ext only
+            $ext = $request->file('cover_image')->getClientOriginalExtension();
+
+            //Filename to store
+            $fileNameToStore = $fileName.'_'.time().'.'.$ext;  //prevents override of img with same name as previous img
+            //Upload img
+            $path = $request->file('cover_image')->storeAs('public/cover_image', $fileNameToStore);
+        } 
 
         //update post
         $post = Post::find($id); //since updating, just search
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->cover_image = $fileNameToStore;
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Updated');
@@ -114,8 +168,19 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $p = Post::find($id);
-        $p->delete();
 
+        //check for correct user so you cannot manually add edit to url
+        if(auth()->user()->id != $p->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+        
+        //ck if there is an img to delete
+        if($p->cover_image != 'noimage.jpg') {
+            //Delete img
+            Storage::delete('public/cover_image/'.$p->cover_image);
+        }
+
+        $p->delete();
         return redirect('/posts')->with('success', 'Post Removed');
     }
 }
